+++ src/mpl_stereo/example_data.py
@@ -12,6 +12,13 @@ def trefoil(i: int = 0, n_points: int = 100, n_steps: int = 10):
 
 def sun_left_right():
     currdir = Path(__file__).parent.resolve()
+    # Get only the red channel for a grayscale image
     sun_left_data = mpimg.imread(currdir / 'sun_left.png')[:, :, 0]
     sun_right_data = mpimg.imread(currdir / 'sun_right.png')[:, :, 0]
+    return sun_left_data, sun_right_data
+
+def church_left_right():
+    currdir = Path(__file__).parent.resolve()
+    sun_left_data = mpimg.imread(currdir / 'church_left.jpg')
+    sun_right_data = mpimg.imread(currdir / 'church_right.jpg')
     return sun_left_data, sun_right_data